more reducers using vanilla JS
