+++ src/reducers/root.js
@@ -1,8 +1,13 @@
 import { combineReducers } from 'redux';
 
 import AutoSuggestReducer from '../header/ducks/auto-suggest/auto-suggest';
+import DataSelectionReducer from '../shared/ducks/data-selection/data-selection';
+import DataSelectionCatalogReducer from '../catalog/ducks/data-selection/data-selection-catalog';
 import ErrorMessageReducer from '../shared/ducks/error-message';
 import PageReducer from '../shared/ducks/page/page';
+import UiReducer from '../shared/ducks/ui/ui';
+import UserReducer from './user';
+import StraatbeeldReducer from '../shared/ducks/straatbeeld/straatbeeld';
 import deprecatedReducer from './deprecated/deprecated-reducer';
 
 // eslint-disable-next-line max-params
@@ -26,36 +31,36 @@ const rootReducer = (oldState, action) => { // eslint-disable-line complexity
   // const ErrorMessageReducer = $window.reducers.ErrorMessageReducer;
   // const MapDetailReducer = $window.reducers.MapDetailReducer;
   // const PanoPreviewReducer = $window.reducers.PanoPreviewReducer;
-  // const UiReducer = $window.reducers.UiReducer;
-  // const DataSelectionReducer = $window.reducers.DataSelectionReducer;
-  // const PageReducer = $window.reducers.PageReducer;
-  // const StraatbeeldReducer = $window.reducers.StraatbeeldReducer;
-  // const UserReducer = $window.reducers.UserReducer;
-  const autoSuggestReducer = AutoSuggestReducer;
-  // const DataSelectionCatalogReducer = $window.reducers.DataSelectionCatalogReducer;
+  // const UiReducer = $window.reducers.UiReducer;                                      // DONE
+  // const DataSelectionReducer = $window.reducers.DataSelectionReducer;                // DONE
+  // const PageReducer = $window.reducers.PageReducer;                                  // DONE
+  // const StraatbeeldReducer = $window.reducers.StraatbeeldReducer;                    // DONE
+  // const UserReducer = $window.reducers.UserReducer;                                  // DONE
+  // const autoSuggestReducer = $window.reducers.AutoSuggestReducer;                    // DONE
+  // const DataSelectionCatalogReducer = $window.reducers.DataSelectionCatalogReducer;  // DONE
   const newRootReducer = combineReducers({
-    //     dataSelection: DataSelectionReducer,
+    dataSelection: DataSelectionReducer,
     page: PageReducer,
     error: ErrorMessageReducer,
     //     map: MapReducer,
     //     mapDetail: MapDetailReducer,
     //     pano: PanoPreviewReducer,
-    //     straatbeeld: StraatbeeldReducer,
-    //     ui: UiReducer,
-    //     user: UserReducer,
+    straatbeeld: StraatbeeldReducer,
+    ui: UiReducer,
+    user: UserReducer,
     //     mapLayers,
-    autoSuggest: autoSuggestReducer
-    //     catalogFilters: DataSelectionCatalogReducer
+    autoSuggest: AutoSuggestReducer,
+    catalogFilters: DataSelectionCatalogReducer
   });
   const filteredState = {
-    //     dataSelection: deprecatedState.dataSelection,
+    dataSelection: deprecatedState.dataSelection,
     page: deprecatedState.page,
     //     map: deprecatedState.map,
     //     mapDetail: deprecatedState.mapDetail,
-    //     straatbeeld: deprecatedState.straatbeeld,
-    //     ui: deprecatedState.ui,
-    //     user: deprecatedState.user,
-    //
+    straatbeeld: deprecatedState.straatbeeld,
+    ui: deprecatedState.ui,
+    user: deprecatedState.user,
+
     //     // Using oldState instead of chaining deprecatedState from
     //     // other reducer for the following fields.
     //     // This is because these fields do not recide in the URL state,
@@ -64,20 +69,17 @@ const rootReducer = (oldState, action) => { // eslint-disable-line complexity
     error: oldState.error,
     //     pano: oldState.pano,
     //     mapLayers: oldState.mapLayers,
-    autoSuggest: oldState.autoSuggest
-    //     catalogFilters: oldState.catalogFilters
+    autoSuggest: oldState.autoSuggest,
+    catalogFilters: oldState.catalogFilters
   };
-  //
+
   // // Combine old and new reducer states
   const newState = {
     ...deprecatedState,
     ...newRootReducer(filteredState, action)
-    // ...oldState,
-    // ...newRootReducer(filteredState, action)
   };
 
   return newState;
-  // return oldState;
 };
 
 export default rootReducer;
