+++ src/test-index.js
@@ -42,10 +42,7 @@ import '../modules/atlas/services/piwik/piwik.factory';
 import '../modules/atlas/services/piwik/piwik.run';
 import '../modules/atlas/services/redux/middleware/context-middleware.factory';
 import '../modules/atlas/services/redux/middleware/state-to-url-middleware.factory';
-import '../modules/atlas/services/redux/reducers/data-selection-reducers.factory';
-import '../modules/atlas/services/redux/reducers/filters-reducers.factory';
 import '../modules/atlas/services/redux/reducers/search-reducers.factory';
-import '../modules/atlas/services/redux/reducers/straatbeeld-reducers.factory';
 import '../modules/atlas/services/redux/reducers/url-reducers.factory';
 import '../modules/atlas/services/redux/store.run';
 import '../modules/atlas/services/routing/state-to-url.factory';
@@ -379,9 +376,6 @@ import '../modules/atlas/services/redux/middleware/state-to-url-middleware.facto
 import '../modules/atlas/services/redux/deprecated-reducer.factory.test';
 import '../modules/atlas/services/redux/reducers/url-reducers.factory.test';
 import '../modules/atlas/services/redux/reducers/search-reducers.factory.test';
-import '../modules/atlas/services/redux/reducers/data-selection-reducers.factory.test';
-import '../modules/atlas/services/redux/reducers/straatbeeld-reducers.factory.test';
-import '../modules/atlas/services/redux/reducers/filters-reducers.factory.test';
 import '../modules/atlas/services/piwik/piwik.factory.test';
 import '../modules/atlas/services/routing/url-to-state.factory.test';
 import '../modules/atlas/services/routing/state-url-conversion.factory.test';
